Move FFI code into ffi module
