+++ libindy-crypto/src/ffi/mod.rs
@@ -1,69 +1,5 @@
+#[macro_use]
+mod ctypes;
 pub mod cl;
 pub mod bls;
 pub mod logger;
-
-#[derive(Debug, PartialEq, Copy, Clone)]
-#[repr(usize)]
-pub enum ErrorCode
-{
-    Success = 0,
-
-    // Common errors
-
-    // Caller passed invalid value as param 1 (null, invalid json and etc..)
-    CommonInvalidParam1 = 100,
-
-    // Caller passed invalid value as param 2 (null, invalid json and etc..)
-    CommonInvalidParam2 = 101,
-
-    // Caller passed invalid value as param 3 (null, invalid json and etc..)
-    CommonInvalidParam3 = 102,
-
-    // Caller passed invalid value as param 4 (null, invalid json and etc..)
-    CommonInvalidParam4 = 103,
-
-    // Caller passed invalid value as param 5 (null, invalid json and etc..)
-    CommonInvalidParam5 = 104,
-
-    // Caller passed invalid value as param 6 (null, invalid json and etc..)
-    CommonInvalidParam6 = 105,
-
-    // Caller passed invalid value as param 7 (null, invalid json and etc..)
-    CommonInvalidParam7 = 106,
-
-    // Caller passed invalid value as param 8 (null, invalid json and etc..)
-    CommonInvalidParam8 = 107,
-
-    // Caller passed invalid value as param 9 (null, invalid json and etc..)
-    CommonInvalidParam9 = 108,
-
-    // Caller passed invalid value as param 10 (null, invalid json and etc..)
-    CommonInvalidParam10 = 109,
-
-    // Caller passed invalid value as param 11 (null, invalid json and etc..)
-    CommonInvalidParam11 = 110,
-
-    // Caller passed invalid value as param 11 (null, invalid json and etc..)
-    CommonInvalidParam12 = 111,
-
-    // Invalid library state was detected in runtime. It signals library bug
-    CommonInvalidState = 112,
-
-    // Object (json, config, key, credential and etc...) passed by library caller has invalid structure
-    CommonInvalidStructure = 113,
-
-    // IO Error
-    CommonIOError = 114,
-
-    // Trying to issue non-revocation credential with full anoncreds revocation accumulator
-    AnoncredsRevocationAccumulatorIsFull = 115,
-
-    // Invalid revocation accumulator index
-    AnoncredsInvalidRevocationAccumulatorIndex = 116,
-
-    // Credential revoked
-    AnoncredsCredentialRevoked = 117,
-
-    // Proof rejected
-    AnoncredsProofRejected = 118,
-}
