+++ libindy-crypto/src/utils/mod.rs
@@ -1,5 +1,3 @@
-#[macro_use]
-pub mod ctypes;
 pub mod commitment;
 pub mod rsa;
 #[macro_use]
